pyproject.toml
[project]
name = "flash-hob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nvidia-cutlass-dsl>=4.2.1",
    "ruff>=0.14.2",
    "torch>=2.9.0",
    "ty>=0.0.1a24",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "fhob"
module-root = ""


[tool.uv]
package = true
python-preference = "only-managed"
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'aarch64'", # Grace-Blackwell
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]
index-strategy = "unsafe-best-match"

[project.scripts]
main = "fhob.main:main"

[tool.uv.extra-build-dependencies]
flash-attn = [
    { requirement = "torch", match-runtime = true },
    "packaging",
    "wheel",
]

[tool.uv.extra-build-variables] # To build on GB10
flash-attn = { CUDA_HOME = "/usr/local/cuda", FLASH_ATTENTION_TRITON_AMD_ENABLE = "TRUE", MAX_JOBS = "10", FLASH_ATTN_CUDA_ARCHS = "120" }

[tool.uv.sources]
torch = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
